+++ K8sUtils/private/rollbackAndWarn.ps1
@@ -27,7 +27,7 @@ function rollbackAndWarn {
             $content = Get-Content $errFile -Raw
             if ($exit -ne 0 -and ($content -like '*Error: release has no 0 version*' -or $content -like '*Error: release: not found*')) {
                 Write-Verbose "Last exit code on rollback was $exit."
-                Write-Status "helm rollback failed, trying uninstall" -LogLevel Error -Char '-'
+                Write-Status "Helm rollback failed, trying uninstall" -LogLevel Error
                 helm uninstall $ReleaseName 2>&1 | Write-MyHost
             }
             Remove-Item $errFile -ErrorAction SilentlyContinue
@@ -43,7 +43,7 @@ function rollbackAndWarn {
         return [RollbackStatus]::DeployedOk
     } catch {
         Write-Warning "Caught error rolling back in catch"
-        Write-Error "$_`n$($_.ScriptStackTrace)"
+        Write-Warning "$_`n$($_.ScriptStackTrace)"
         return [RollbackStatus]::HelmStatusFailed
     }
 }