+++ K8sUtils/public/Get-DeploymentStatus.ps1
@@ -48,26 +48,26 @@ function Get-DeploymentStatus {
 
     # get the deployment to get the replica count, loop since it may not be ready yet
     $replicas = $null
-    $uid = $null
     for ( $i = 0; $i -lt 10 -and $null -eq $replicas; $i++) {
         # todo check to see if it exists, or don't use jsonpath since items[0] can fail
-        Write-Verbose "kubectl get deploy --namespace $Namespace -l $Selector -o jsonpath='{.items}'"
-        $deployments = kubectl get deploy --namespace $Namespace -l $Selector -o jsonpath='{.items}' | ConvertFrom-Json -Depth 20
+        Write-Verbose "kubectl get deploy --namespace $Namespace -l $Selector -o jsonpath='{.items}' | ConvertFrom-Json -Depth 20 -AsHashtable"
+        $deployments = kubectl get deploy --namespace $Namespace -l $Selector -o jsonpath='{.items}' | ConvertFrom-Json -Depth 20 -AsHashtable
         if (!$deployments) {
             Write-Warning "No items from kubectl get deploy -l $Selector. Trying again in 1 second."
         } else {
             $deployment = $deployments | Select-Object -First 1 # handle as hash table
             $replicas = $deployment.spec.replicas
-            $uid = $deployment.metadata.uid
             $revision = $deployment.metadata.annotations.'deployment.kubernetes.io/revision'
         }
         Start-Sleep -Seconds 1
     }
-    if ($LASTEXITCODE -ne 0 || $null -eq $replicas) {
-        throw "No data from kubectl get deploy -l $Selector"
+    if ($LASTEXITCODE -ne 0 -or $null -eq $replicas) {
+        throw "Check chart name. No data from kubectl get deploy -l $Selector"
     }
 
     # get the current replicaSet's for hash to get pods in this deployment
+    # && doesn't seems to work to get owner and revision in the JSON Path in one call
+    # "jsonpath-as-json={.items[?(@.metadata.ownerReferences[0].name==`"$deploymentName`")]}"
     Write-Verbose "kubectl get rs -l $Selector -o jsonpath={.items[?(@.metadata.annotations.deployment\.kubernetes\.io/revision==`"$revision`")]}"
     $rs = kubectl get rs -l $Selector -o "jsonpath={.items[?(@.metadata.annotations.deployment\.kubernetes\.io/revision==`"$revision`")]}" |
                             ConvertFrom-Json -Depth 20
@@ -77,6 +77,19 @@ function Get-DeploymentStatus {
     }
     $hash = $rs.metadata.labels."pod-template-hash"
     Write-Verbose "rs pod-template-hash is $hash"
+    $rsEvents = Write-K8sEvent -Prefix "ReplicaSet" -ObjectName $rs.metadata.name `
+                                        -Namespace $Namespace `
+                                        -PassThru `
+                                        -LogLevel error `
+                                        -FilterStartupWarnings
+    # $rsEvents = Get-K8sEvent -ObjectName $rs.metadata.name -NoNormal -Namespace $Namespace
+    if ($rsEvents) {
+        Write-Verbose "Rs Events are $($rsEvents | ConvertTo-Json -depth 10)"
+        $ret = [PodStatus]::new("<replica set error>")
+        $ret.Status = [Status]::ConfigError
+        $ret.LastBadEvents = $rsEvents
+        return $ret
+    }
 
     Write-Status "Looking for $replicas pod$($replicas -eq 1 ? '' : 's') with pod-template-hash=$hash" -Length 0 -LogLevel Normal
     $podSelector = "pod-template-hash=$hash"
@@ -88,7 +101,7 @@ function Get-DeploymentStatus {
                          -PollIntervalSec $PollIntervalSec `
                          -LogFileFolder $LogFileFolder
 
-    Write-Verbose "ret is $($ret | out-string)"
+    Write-Verbose "ret for pod is $($ret | out-string)"
 
     return $ret
 }